requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
python-dotenv==0.21.0
python-dotenv-vault==0.0.7
Werkzeug==2.2.2
zipp==3.10.0
